types cleanup

+++ include/rtos/malloc/pmsis_l2_malloc.h
@@ -4,7 +4,7 @@
 #include "pmsis_types.h"
 #include "pmsis_malloc_internal.h"
 
-void *pmsis_l2_malloc(uint32_t size);
+void *pmsis_l2_malloc(size_t size);
 
 void pmsis_l2_malloc_free(void *_chunk, int size);
